Re-organize
Re-organized the entire frontend and updated related imports.
