+++ FitnessApp/src/Assets/Front.jsx
