+++ FitnessApp/src/Pages/Daily/Daily.css
