+++ FitnessApp/src/Pages/Daily/Daily.jsx
