+++ FitnessApp/src/Pages/DataView/BarChart.jsx
