+++ FitnessApp/src/Utils/FuncUtil.jsx
@@ -1,0 +1,6 @@
+export function formatDateToMySQL(date) {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+  return `${year}-${month}-${day}`;
+}
