+++ FitnessApp/src/main.jsx
@@ -2,23 +2,23 @@ import { StrictMode, useContext } from "react";
 import { createRoot } from "react-dom/client";
 //Routing imports
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { Provider } from "./Provider";
+import { Provider } from "./Routing/Provider";
 //Component imports
-import ViewPage from "./DataImpose/ViewPage.jsx";
-import StartPage from "./Landing/StartPage.jsx";
-import LoginPage from "./Login/Login.jsx";
-import Daily from "./Daily/Daily.jsx";
-import Share from "./Share/Share.jsx";
-import Nav from "./NavBar/Nav.jsx";
-import PrivateRoute from "./Routes.jsx";
+import ViewPage from "./Pages/DataView/ViewPage.jsx";
+import StartPage from "./Pages/Landing/StartPage.jsx";
+import LoginPage from "./Pages/Login/Login.jsx";
+import Daily from "./Pages/Daily/Daily.jsx";
+import Share from "./Pages/Share/Share.jsx";
+import Nav from "./Components/NavBar/Nav.jsx";
+import PrivateRoute from "./Routing/Routes.jsx";
 //CSS imports
 import "./Site.css";
-import "./Landing/StartPage.css";
-import "./DataImpose/ViewPage.css";
-import "./Login/Login.css";
-import "./Daily/Daily.css";
-import "./Share/Share.css";
-import "./NavBar/Nav.css";
+import "./Pages/Landing/StartPage.css";
+import "./Pages/DataView/ViewPage.css";
+import "./Pages/Login/Login.css";
+import "./Pages/Daily/Daily.css";
+import "./Pages/Share/Share.css";
+import "./Components/NavBar/Nav.css";
 
 //
 //The daily page is the input page for the user to input their daily data into the server
